Extract view model to dedicated file
